moved abstract_model to core
