+++ core/src/autogluon/core/models/__init__.py
@@ -1,0 +1,1 @@
+from .abstract.abstract_model import AbstractModel
