+++ tabular/src/autogluon/tabular/models/ensemble/bagged_ensemble_model.py
@@ -16,7 +16,7 @@ from autogluon.core.utils.loaders import load_pkl
 from autogluon.core.utils.savers import save_pkl
 from autogluon.core.utils.utils import _compute_fi_with_stddev
 
-from ..abstract.abstract_model import AbstractModel
+from autogluon.core.models import AbstractModel
 
 logger = logging.getLogger(__name__)
