+++ tabular/src/autogluon/tabular/models/ensemble/greedy_weighted_ensemble_model.py
@@ -3,8 +3,8 @@ import logging
 from autogluon.core.constants import MULTICLASS
 from autogluon.core.features.feature_metadata import S_STACK
 
-from ..abstract.abstract_model import AbstractModel
-from ..abstract.model_trial import skip_hpo
+from autogluon.core.models import AbstractModel
+from autogluon.core.models.abstract.model_trial import skip_hpo
 from ...tuning.ensemble_selection import EnsembleSelection
 
 logger = logging.getLogger(__name__)
