+++ tabular/src/autogluon/tabular/models/ensemble/stacker_ensemble_model.py
@@ -9,7 +9,7 @@ from autogluon.core.utils.utils import generate_kfold
 from autogluon.core.constants import MULTICLASS
 from autogluon.core.features.feature_metadata import FeatureMetadata, R_FLOAT, S_STACK
 
-from ..abstract.abstract_model import AbstractModel
+from autogluon.core.models import AbstractModel
 from .bagged_ensemble_model import BaggedEnsembleModel
 
 logger = logging.getLogger(__name__)
