+++ tabular/src/autogluon/tabular/models/knn/knn_model.py
@@ -11,8 +11,8 @@ from autogluon.core.utils.exceptions import NotEnoughMemoryError
 from autogluon.core.features.feature_metadata import R_CATEGORY, R_OBJECT, S_TEXT_NGRAM, S_TEXT_SPECIAL, S_DATETIME_AS_INT
 
 from .knn_utils import FAISSNeighborsClassifier, FAISSNeighborsRegressor
-from ..abstract.model_trial import skip_hpo
-from ..abstract.abstract_model import AbstractModel
+from autogluon.core.models.abstract.model_trial import skip_hpo
+from autogluon.core.models import AbstractModel
 
 logger = logging.getLogger(__name__)
