+++ tabular/src/autogluon/tabular/models/rf/rf_model.py
@@ -12,8 +12,8 @@ from autogluon.core.constants import MULTICLASS, REGRESSION
 from autogluon.core.utils.exceptions import NotEnoughMemoryError, TimeLimitExceeded
 from autogluon.core.features.feature_metadata import R_OBJECT
 
-from ..abstract.model_trial import skip_hpo
-from ..abstract.abstract_model import AbstractModel
+from autogluon.core.models.abstract.model_trial import skip_hpo
+from autogluon.core.models import AbstractModel
 from ...features.generators import LabelEncoderFeatureGenerator
 
 logger = logging.getLogger(__name__)
